+++ test/functional/find-options/empty-properties/entity/Post.ts
@@ -1,0 +1,15 @@
+import {Column, Entity, PrimaryGeneratedColumn} from "../../../../../src";
+
+@Entity()
+export class Post {
+
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column()
+    title: string;
+
+    @Column()
+    text: string;
+
+}